Add new 🎲 and 🎯 emojis
